Update Command Pattern

+++ DesignPattern/src/BehavioralPatterns/Command/Addition.java
@@ -3,8 +3,8 @@ package BehavioralPatterns.Command;
 public class Addition implements Command {
 	
 	@Override
-	public String execute(int a, int b) {
-		int result = a + b;
+	public String execute(double a, double b) {
+		double result = a + b;
 		return "Addition results: " + result;
 	};
