+++ DesignPattern/src/BehavioralPatterns/Command/Calculator.java
@@ -2,7 +2,7 @@ package BehavioralPatterns.Command;
 
 public class Calculator {
 	
-	public String calculate(Command command, int a, int b) {
+	public String calculate(Command command, double a, double b) {
 		return command.execute(a, b);
 	}
