+++ DesignPattern/src/BehavioralPatterns/Command/Command.java
@@ -2,6 +2,6 @@ package BehavioralPatterns.Command;
 
 public interface Command {
 	
-	public String execute(int a, int b);
+	public String execute(double a, double b);
 
 }
