+++ DesignPattern/src/BehavioralPatterns/Command/Multiplication.java
@@ -1,10 +1,10 @@
 package BehavioralPatterns.Command;
 
-public class Multiplication implements Command  {
+public class Multiplication implements Command {
 	
 	@Override
-	public String execute(int a, int b) {
-		int result = a * b;
+	public String execute(double a, double b) {
+		double result = a * b;
 		return "Multiplication results: " + result;
 	};
 }
